pyproject.toml
[tool.poetry]
name = "vidypy"
version = "0.1.0"
description = ""
authors = ["Iván Salazar <ivangio.salazar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.3"
yt-dlp = "^2023.6.22"
validators = "^0.20.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
autoimport = "^1.3.3"
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
